Move the regression tests out of the benchmark module
test_sfsolve.py sets pytestmark = pytest.mark.benchmark at module scope, so
the eight regression tests added in this branch were all swept into the
CodSpeed run, taking the instrumented set from 5 tests to 13. Several of
them only exercise error paths, which is meaningless as a performance
measurement, and it lengthens a job that already takes 25 minutes while
adding permanent noise to the tracked benchmark set.

Move them to test_regressions.py without the benchmark mark. All 13 still
run in the normal pytest job; the instrumented set is back to the same five
cases as main (with test_invalid_excitation_fraction_ignored renamed to
_reported, since that behaviour changed earlier in this branch).
